+++ pangea-sdk/service/ai_guard/integration_test.go
@@ -20,19 +20,39 @@ func TestTextGuard(t *testing.T) {
 
 	client := ai_guard.New(pangeatesting.IntegrationConfig(t, testingEnvironment))
 
-	input := &ai_guard.TextGuardRequest{Text: "hello world", Recipe: "pangea_prompt_guard"}
+	input := &ai_guard.TextGuardRequest{Text: "what was pangea?"}
 	out, err := client.GuardText(ctx, input)
 	assert.NoError(t, err)
 	assert.NotNil(t, out.Result)
-	assert.NotNil(t, out.Result.Prompt)
+	assert.NotNil(t, out.Result.PromptText)
 	assert.False(t, out.Result.Detectors.PromptInjection.Detected)
 	assert.False(t, out.Result.Detectors.PiiEntity.Detected)
-	assert.False(t, out.Result.Detectors.MaliciousEntity.Detected)
+	if out.Result.Detectors.MaliciousEntity != nil {
+		assert.False(t, out.Result.Detectors.MaliciousEntity.Detected)
+	}
 
-	input = &ai_guard.TextGuardRequest{Text: "security@pangea.cloud", Recipe: "pangea_prompt_guard"}
+	input = &ai_guard.TextGuardRequest{Text: "security@pangea.cloud"}
 	out, err = client.GuardText(ctx, input)
 	assert.NoError(t, err)
 	assert.NotNil(t, out.Result)
-	assert.NotNil(t, out.Result.Prompt)
+	assert.NotNil(t, out.Result.PromptText)
 	assert.True(t, out.Result.Detectors.PiiEntity.Detected)
+}
+
+func TestTextGuard_Messages(t *testing.T) {
+	ctx, cancelFn := context.WithTimeout(context.Background(), 5*time.Second)
+	defer cancelFn()
+
+	client := ai_guard.New(pangeatesting.IntegrationConfig(t, testingEnvironment))
+
+	input := &ai_guard.TextGuardRequest{Messages: []map[string]interface{}{
+		{
+			"role":    "user",
+			"content": "what was pangea?",
+		},
+	}}
+	out, err := client.GuardText(ctx, input)
+	assert.NoError(t, err)
+	assert.NotNil(t, out.Result)
+	assert.NotNil(t, out.Result.PromptMessages)
 }