+++ CHANGELOG.md
@@ -7,10 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- AI Guard: support for structured input.
+- AI Guard: more detectors.
+
 ### Changed
 
 - Deprecated `Config.Enviroment` (typo) in favor of `Config.Environment`.
 - Clarified what `Config.Environment` affects.
+- AI Guard: replaced `redacted` with `action`.
 
 ## 4.4.0-beta.2 - 2025-01-21